Finish rna-transcription for JS track
This was pretty easy. I used an object with keys corresponding to each nucleotide and the value being their complement.

Then I wrote a one liner that splits the string into an array, maps over and returns the complement based on the key and then joins the array back into a string.